@@ -6,56 +6,54 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 
-namespace SemanticKernel.FunctionCalling
+namespace SKFunctionCalling;
+public class DeleteModel : PageModel
 {
-    public class DeleteModel : PageModel
+    private readonly ApplicationDbContext _context;
+
+    public DeleteModel(ApplicationDbContext context)
     {
-        private readonly ApplicationDbContext _context;
+        _context = context;
+    }
 
-        public DeleteModel(ApplicationDbContext context)
+    [BindProperty]
+    public Student Student { get; set; } = default!;
+
+    public async Task<IActionResult> OnGetAsync(int? id)
+    {
+        if (id == null)
         {
-            _context = context;
+            return NotFound();
         }
 
-        [BindProperty]
-        public Student Student { get; set; } = default!;
+        var student = await _context.Students.FirstOrDefaultAsync(m => m.StudentId == id);
 
-        public async Task<IActionResult> OnGetAsync(int? id)
+        if (student == null)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            return NotFound();
+        }
+        else
+        {
+            Student = student;
+        }
+        return Page();
+    }
 
-            var student = await _context.Students.FirstOrDefaultAsync(m => m.StudentId == id);
-
-            if (student == null)
-            {
-                return NotFound();
-            }
-            else
-            {
-                Student = student;
-            }
-            return Page();
+    public async Task<IActionResult> OnPostAsync(int? id)
+    {
+        if (id == null)
+        {
+            return NotFound();
         }
 
-        public async Task<IActionResult> OnPostAsync(int? id)
+        var student = await _context.Students.FindAsync(id);
+        if (student != null)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var student = await _context.Students.FindAsync(id);
-            if (student != null)
-            {
-                Student = student;
-                _context.Students.Remove(Student);
-                await _context.SaveChangesAsync();
-            }
-
-            return RedirectToPage("./Index");
+            Student = student;
+            _context.Students.Remove(Student);
+            await _context.SaveChangesAsync();
         }
+
+        return RedirectToPage("./Index");
     }
 }
