@@ -1,9 +1,6 @@
-﻿using System.Globalization;
-using CsvHelper;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 
-namespace SemanticKernel.FunctionCalling;
+namespace SKFunctionCalling;
 
 public class ApplicationDbContext : DbContext
 {
@@ -13,5 +10,21 @@ public class ApplicationDbContext : DbContext
         DBInitializer.Initialize(this);
     }
     public DbSet<Student> Students { get; set; }
-   
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        base.OnModelCreating(modelBuilder);
+
+        // Set case-insensitive comparison for all string properties
+        foreach (var entityType in modelBuilder.Model.GetEntityTypes())
+        {
+            foreach (var property in entityType.GetProperties())
+            {
+                if (property.ClrType == typeof(string))
+                {
+                    property.SetCollation("NOCASE");
+                }
+            }
+        }
+    }
 }
