@@ -1,8 +1,8 @@
-using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace SemanticKernel.FunctionCalling;
-public class Student {
+namespace SKFunctionCalling;
+public class Student
+{
    public int StudentId { get; set; }
 
    [Display(Name = "First Name")]
@@ -15,8 +15,12 @@ public class Student {
 
    [Required]
    public string? School { get; set; }
- 
-   public override string ToString() {
+
+   [Required]
+   public int Age { get; set; }
+
+   public override string ToString()
+   {
       return $"Student ID: {StudentId}, First Name: {FirstName}, Last Name: {LastName}, School: {School}";
    }
 }
