@@ -1,13 +1,11 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 
-namespace SemanticKernel.FunctionCalling;
+namespace SKFunctionCalling;
 
-public class Utils {
-  public static string GetConfigValue(string config) {
+public class Utils
+{
+  public static string GetConfigValue(string config)
+  {
     IConfigurationBuilder builder = new ConfigurationBuilder();
     if (System.IO.File.Exists("appsettings.json"))
       builder.AddJsonFile("appsettings.json", false, true);
@@ -16,8 +14,9 @@ public class Utils {
     IConfigurationRoot root = builder.Build();
     return root[config]!;
   }
- 
-  public static ApplicationDbContext GetDbContext() {
+
+  public static ApplicationDbContext GetDbContext()
+  {
     var optionsBuilder = new DbContextOptionsBuilder<ApplicationDbContext>();
     var connStr = Utils.GetConfigValue("ConnectionStrings:DefaultConnection");
     optionsBuilder.UseSqlite(connStr);
