@@ -1,5 +1,5 @@
 ﻿@page
-@model IndexModel
+@model SKFunctionCalling.IndexModel
 
 <div class="text-center">
     <h4 class="display-6">Be informed about students:</h4>
