@@ -4,16 +4,18 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>@ViewData["Title"]</title>
+    <title>@ViewData["Title"] - Students AI</title>
+
     <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
     <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
+    <link rel="stylesheet" href="~/razor_demo.styles.css" asp-append-version="true" />
 </head>
 
 <body>
     <header>
         <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3">
             <div class="container">
-                <a class="navbar-brand" asp-area="" asp-page="/Index">SK Students AI</a>
+                <a class="navbar-brand" asp-area="" asp-page="/Index">Students AI</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target=".navbar-collapse"
                     aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
@@ -27,7 +29,6 @@
                             <a class="nav-link text-dark" asp-area="" asp-page="/StudentPages/Index">Students</a>
                         </li>
                     </ul>
-                    <partial name="_LoginPartial" />
                 </div>
             </div>
         </nav>
@@ -40,6 +41,7 @@
 
     <footer class="border-top footer text-muted">
         <div class="container">
+            &copy; 2025 - razor_demo - <a asp-area="" asp-page="/Privacy">Privacy</a>
         </div>
     </footer>
 
