@@ -1,5 +1,5 @@
 @page
-@model SemanticKernel.FunctionCalling.DeleteModel
+@model SKFunctionCalling.DeleteModel
 
 @{
     ViewData["Title"] = "Delete";
@@ -31,7 +31,7 @@
             @Html.DisplayFor(model => model.Student.School)
         </dd>
     </dl>
-    
+
     <form method="post">
         <input type="hidden" asp-for="Student.StudentId" />
         <input type="submit" value="Delete" class="btn btn-danger" /> |
