@@ -1,5 +1,5 @@
 @page
-@model SemanticKernel.FunctionCalling.DetailsModel
+@model SKFunctionCalling.DetailsModel
 
 @{
     ViewData["Title"] = "Details";
