@@ -1,5 +1,5 @@
 @page
-@model SemanticKernel.FunctionCalling.EditModel
+@model SKFunctionCalling.EditModel
 
 @{
     ViewData["Title"] = "Edit";
@@ -41,5 +41,7 @@
 </div>
 
 @section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+    @{
+        await Html.RenderPartialAsync("_ValidationScriptsPartial");
+    }
 }
