@@ -1,4 +1,3 @@
-﻿@using Microsoft.AspNetCore.Identity
-@using SemanticKernel.FunctionCalling
-@namespace SemanticKernel.FunctionCalling
+﻿@using sk_function_calling
+@namespace sk_function_calling.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
